Initial code to convert routes into Rails
Lots of find and replaces and changing how routes are defined in
Sinatra vs rails. For the most part it was just taking
`get '/uri_path/route_name' do` to `def route_name`.

Changed `@env` to `request.env`

`generate_login_ticket` requests a client string to be passed in.
Originally, the function found it through the request.

Since aren't using R18n, I just manually added the en version of
the strings into the routes.

I updated the module paths and names with the ones that could be
found in rubycas-server-core.